Update 9.5-9.6 Генерация перестановок.cpp
